refactor(apps): rewrite current user logic
