+++ packages/app-base/app.vue
@@ -2,7 +2,7 @@
   import { changeLocale } from '@formkit/vue';
   import { useAuthStore } from '@tituskirch/app-base/stores/auth';
   import { useFeatureStore } from '@tituskirch/app-base/stores/feature';
-  const { currentUser } = useNewCurrentUser();
+  const { currentUser } = useCurrentUser();
 
   // head
   const route = useRoute();
