+++ packages/app-base/components/auth/ResetPasswordForm.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-  const { fetchCurrentUser } = useNewCurrentUser();
+  const { fetchCurrentUser } = useCurrentUser();
 
   // get data from route
   const route = useRoute();
