+++ packages/app-base/components/auth/TwoFactorChallengeForm.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-  const { fetchCurrentUser } = useNewCurrentUser();
+  const { fetchCurrentUser } = useCurrentUser();
 
   withDefaults(
     defineProps<{
